Moved SmiXetterAccessMap to shared src folder (#1310)
